Fix concurrent access to the cells set.
